requirements.txt
django==3.1.3
djangorestframework==3.12.2
psycopg2==2.8.6
Pillow==8.0.1
drf-yasg==1.20.0
djangorestframework-jsonapi==4.0.0
django-filter==2.4.0
django-geojson-serializer==0.3